+++ src/pages/ProjectPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import ProjectPageContent from "../components/ProjectPage/ProjectPageContent";
+import { Outlet } from "react-router-dom";
 
 export default function ProjectPage() {
-  return <ProjectPageContent isAdmin={true} />;
+  return <Outlet />;
 }
