+++ src/router/Routes.jsx
@@ -5,7 +5,7 @@ import AboutPage from '@pages/AboutPage';
 import ProjectPage from '@pages/ProjectPage';
 import RecruitPage from '@pages/RecruitPage';
 import LoginPage from '@pages/LoginPage';
-import ProjectPageContent from '../components/ProjectPage/ProjectPageContent';
+import ProjectPageLayout from '../components/ProjectPage/ProjectPageLayout';
 import { MainLayout } from '@layouts';
 
 export default function Routes() {
@@ -17,12 +17,9 @@ export default function Routes() {
         <Route path="about" element={<AboutPage />} />
         <Route path="project" element={<ProjectPage />}>
           {/* 사용자 모드 */}
-          <Route index element={<ProjectPageContent isAdmin={false} />} />
+          <Route index element={<ProjectPageLayout isAdmin={false} />} />
           {/* 관리자 모드 */}
-          <Route
-            path="admin"
-            element={<ProjectPageContent isAdmin={true} />} 
-          />
+          <Route path="admin" element={<ProjectPageLayout isAdmin={true} />} />
         </Route>
         <Route path="recruit" element={<RecruitPage />} />
         <Route path="login" element={<LoginPage />} />
